Refactor: Extra Category init into A factory method
